backend/requirements.txt
fastapi
uvicorn
scikit-learn
PyPDF2
python-docx
python-multipart
sentence-transformers
google-generativeai
python-dotenv